requirements.txt
#############################
# Minimal runtime dependencies
#############################
Flask==3.1.1
Flask-Login==0.6.3
python-dotenv==1.1.1
psycopg2-binary==2.9.10
Alembic==1.13.2
gunicorn==23.0.0
requests==2.32.4

#############################
# AI / LLM features
#############################
langchain==0.3.26
langchain-core>=0.3.66,<1.0.0  # Must match langchain version compatibility
langchain-openai==0.3.29
langchain-community==0.3.26
chromadb==1.0.15
openai==1.99.9
tiktoken==0.9.0
fpdf==1.7.2
python-docx==1.1.2  # DOCX export for petição
faiss-cpu>=1.9.0  # Vector similarity search for ementas
sentence-transformers>=2.2.0  # Embeddings for ementas search
gensim>=4.0.0  # Word2Vec embeddings for legal BiLSTM

#############################
# Optional document/NLP (uncomment if needed)
#############################
pdfplumber==0.11.7
pdfminer.six==20250506
pytesseract==0.3.13
spacy==3.8.7
moviepy==1.0.3
pt_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.8.0/pt_core_news_sm-3.8.0-py3-none-any.whl

#############################
# Dev / tooling (create requirements-dev.txt ideally)
#############################
# pytest==8.4.1
# mypy==1.17.1

#############################
# Notes
# 1. Keep this list lean for faster builds.
# 2. Add only what is directly imported.
# 3. Freeze exact versions for reproducible Docker builds.
#############################
